move s3e to wordembedding_utils.py
